+++ src/ui_printer.cpp
@@ -741,12 +741,14 @@ void update_printer_ui() {
         lv_label_set_text(g_remain_label, rt);
     }
 
-    lv_label_set_text_fmt(g_nozzle_label, "%s %.0f/%.0f\xC2\xB0" "C", T("dash.nozzle"), s.nozzle_temp, s.nozzle_target);
-    lv_label_set_text_fmt(g_bed_label, "%s %.0f/%.0f\xC2\xB0" "C", T("dash.bed"), s.bed_temp, s.bed_target);
+    // Bambu truncates temperatures (23.9 shows as 23), so we do too - keeps our
+    // readout matching Bambu Studio/Handy instead of rounding a degree higher.
+    lv_label_set_text_fmt(g_nozzle_label, "%s %d/%d\xC2\xB0" "C", T("dash.nozzle"), (int)s.nozzle_temp, (int)s.nozzle_target);
+    lv_label_set_text_fmt(g_bed_label, "%s %d/%d\xC2\xB0" "C", T("dash.bed"), (int)s.bed_temp, (int)s.bed_target);
     // Only printers with a real chamber sensor (X1/H2) get a chamber reading; the
     // P1/A1 series report a bogus placeholder, so we leave it blank like Bambu does.
     if (printer_has_chamber_sensor())
-        lv_label_set_text_fmt(g_chamber_label, "%s %.0f\xC2\xB0" "C", T("dash.chamber"), s.chamber_temp);
+        lv_label_set_text_fmt(g_chamber_label, "%s %d\xC2\xB0" "C", T("dash.chamber"), (int)s.chamber_temp);
     else
         lv_label_set_text(g_chamber_label, "");
